+++ CHANGELOG.md
@@ -65,6 +65,7 @@ Breaking changes will be documented in this changelog file for each version.
 * github-336: Add documentation for flowman-kernel
 * github-366: Feature parity between Flowman shell and Flowman remote shell
 * github-365: Implement saving mappings in Flowman Kernel/client
+* github-367: Create integration test for "quickstart" archetype
 
 ### Breaking changes